pyproject.toml
[project]
name = "py-suiteql"
version = "0.1.0"
description = "Execute SuiteQL queries against NetSuite from Python."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "certifi>=2025.11.12",
    "charset-normalizer>=3.4.1",
    "idna>=3.10",
    "oauthlib>=3.2.2",
    "prompt-toolkit>=3.0",
    "pygments>=2.18",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
    "rich>=13.0",
    "sqlparse>=0.5",
    "urllib3>=2.3.0",
]
